delete 10 and move ConsumerFullInfo.java to resps
