+++ src/main/java/redis/clients/jedis/Jedis.java
@@ -4469,8 +4469,7 @@ public class Jedis implements ServerCommands, DatabaseCommands, JedisCommands, J
   @Override
   public StreamFullInfo xinfoStreamFull(byte[] key) {
     checkIsInMultiOrPipeline();
-    return connection.executeCommand(commandObjects.xinfoStreamFull(key, 10));
-
+    return connection.executeCommand(commandObjects.xinfoStreamFull(key));
   }
 
   @Override
