+++ src/main/java/redis/clients/jedis/Pipeline.java
@@ -2597,7 +2597,7 @@ public class Pipeline extends Queable  implements PipelineCommands, PipelineBina
 
   @Override
   public Response<StreamFullInfo> xinfoStreamFull(String key) {
-    return appendCommand(commandObjects.xinfoStreamFull(key, 10));
+    return appendCommand(commandObjects.xinfoStreamFull(key));
   }
 
   @Override
