+++ src/main/java/redis/clients/jedis/TransactionBase.java
@@ -1415,7 +1415,7 @@ public abstract class TransactionBase extends Queable implements PipelineCommand
 
   @Override
   public Response<StreamFullInfo> xinfoStreamFull(String key) {
-    return appendCommand(commandObjects.xinfoStreamFull(key, 10));
+    return appendCommand(commandObjects.xinfoStreamFull(key));
   }
 
   @Override
